+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file. The format
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 3.0.6
+
+### Changed
+- Remove the menu screen when no menu modal is open.
+- Update the mobile nav menu layout.
+
+
 ## 3.0.5
 
 ### Changed